8th

+++ Wiem-app/routes/web.php
@@ -15,22 +15,10 @@ use App\Http\Controllers\CvController;
 |
 */
 
-//Integration python test
-// Route::get('/class/{id}', [UploadController::class,'analyse'])->name('analyse');
-
-// Route::post('/cvs', [App\Http\Controllers\CvController::class, 'store'])->name('store');
 
 
-
-
-// Route::post('/analyse/{id}', [UploadController::class,'analyse'])->name('analyse');
-
-
-
-
-
-Route::get('/profile', [App\Http\Controllers\HomeController::class, 'showChangePasswordGet'])->name('changePasswordGet');
-Route::post('/profile', [App\Http\Controllers\HomeController::class, 'changePasswordPost'])->name('changePasswordPost');
+Route::get('/profile', [App\Http\Controllers\HomeController::class, 'showChangePasswordGet'])->middleware(['auth'])->name('changePasswordGet');
+Route::post('/profile', [App\Http\Controllers\HomeController::class, 'changePasswordPost'])->middleware(['auth'])->name('changePasswordPost');
 
 
 Route::get('/', function () {
@@ -79,47 +67,47 @@ Route::get('/dashboard', function () {
 
 
 
-Route::get('/search/',[App\Http\Controllers\CvController::class, 'search'])->name('search');
+Route::get('/search/',[App\Http\Controllers\CvController::class, 'search'])->middleware(['auth'])->name('search');
 
 
-Route::get('/cvs/create', [App\Http\Controllers\CvController::class, 'create'])->name('create');
-Route::post('/cvs', [App\Http\Controllers\CvController::class, 'store'])->name('store');
-Route::get('/cvs', [App\Http\Controllers\CvController::class, 'index'])->name('index');
-Route::get('/cvs/{id}/edit', [App\Http\Controllers\CvController::class, 'edit'])->name('edit');
-Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'update'])->name('update');
-Route::delete('/cvs/{id}', [App\Http\Controllers\CvController::class, 'destroy'])->name('destroy');
+Route::get('/cvs/create', [App\Http\Controllers\CvController::class, 'create'])->middleware(['auth'])->name('create');
+Route::post('/cvs', [App\Http\Controllers\CvController::class, 'store'])->middleware(['auth'])->name('store');
+Route::get('/cvs', [App\Http\Controllers\CvController::class, 'index'])->middleware(['auth'])->name('index');
+Route::get('/cvs/{id}/edit', [App\Http\Controllers\CvController::class, 'edit'])->middleware(['auth'])->name('edit');
+Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'update'])->middleware(['auth'])->name('update');
+Route::delete('/cvs/{id}', [App\Http\Controllers\CvController::class, 'destroy'])->middleware(['auth'])->name('destroy');
 
 
 //queries
 //get element of it departement
-Route::get('/it', [App\Http\Controllers\CvController::class, 'getItDep'])->name('getItDep');
+Route::get('/it', [App\Http\Controllers\CvController::class, 'getItDep'])->middleware(['auth'])->name('getItDep');
 //get element of Production departement
-Route::get('/production', [App\Http\Controllers\CvController::class, 'getProdDep'])->name('getProdDep');
+Route::get('/production', [App\Http\Controllers\CvController::class, 'getProdDep'])->middleware(['auth'])->name('getProdDep');
 //get element of Technique departement
-Route::get('/technique', [App\Http\Controllers\CvController::class, 'getTechDep'])->name('getTechDep');
+Route::get('/technique', [App\Http\Controllers\CvController::class, 'getTechDep'])->middleware(['auth'])->name('getTechDep');
 //get element of Administratif departement
-Route::get('/administratif', [App\Http\Controllers\CvController::class, 'getAdDep'])->name('getAdDep');
+Route::get('/administratif', [App\Http\Controllers\CvController::class, 'getAdDep'])->middleware(['auth'])->name('getAdDep');
 //get element of Autres departement
-Route::get('/autres', [App\Http\Controllers\CvController::class, 'getOtherDep'])->name('getOtherDep');
+Route::get('/autres', [App\Http\Controllers\CvController::class, 'getOtherDep'])->middleware(['auth'])->name('getOtherDep');
 
 
 
 //Modif Autre depart
-Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'updateAutre'])->name('updateAutre');
+Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'updateAutre'])->middleware(['auth'])->name('updateAutre');
 
 //Modif Prod depart
-Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'updateprod'])->name('updateprod');
+Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'updateprod'])->middleware(['auth'])->name('updateprod');
 
 
 //Modif Administratif depart
-Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'updateAdministratif'])->name('updateAdministratif');
+Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'updateAdministratif'])->middleware(['auth'])->name('updateAdministratif');
 
 
 //Modif tech depart
-Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'updateTech'])->name('updateTech');
+Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'updateTech'])->middleware(['auth'])->name('updateTech');
 
 //Modif IT depart
-Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'updateIT'])->name('updateIT');
+Route::put('/cvs/{id}', [App\Http\Controllers\CvController::class, 'updateIT'])->middleware(['auth'])->name('updateIT');
 
 
 require __DIR__.'/auth.php';
